src/main.py
import logging
from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=os.getenv('LOG_LEVEL', 'INFO'))
logger = logging.getLogger('LinkedInAutoApply')

def main():
    logger.info('Starting LinkedIn Auto-Apply AI Agent...')
    # TODO: Initialize system components and main loop
    pass

if __name__ == '__main__':
    main()
