requirements.txt
openai>=1.0.0
selenium>=4.15.0
pyautogui>=0.9.54
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
sqlite3
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0
pydantic>=2.0.0
loguru>=0.7.0
